% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_PACKAGES.R
\name{dl_PACKAGES}
\alias{dl_PACKAGES}
\title{Download a CRAN PACKAGES file}
\usage{
dl_PACKAGES(repourl, destfile = "./PACKAGES.rds")
}
\arguments{
\item{repourl}{A CRAN repository URL}

\item{destfile}{the destination file}
}
\description{
Downloads a CRAN packages file from the specified CRAN repo.
}
