% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imports.R
\name{get_imports}
\alias{get_imports}
\title{Get a list of package imports}
\usage{
get_imports(packages, pkgs = readRDS("~/.cranalysis/PACKAGES.rds"))
}
\arguments{
\item{packages}{A list of packages to report on}

\item{pkgs}{The content of the PACKAGES.rds file to report from}
}
\description{
Returns a vector containing the packages that are listed as imports for the input vector of packages
}
