% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cache.R
\name{update_cache}
\alias{update_cache}
\title{update cached CRAN PACKAGES file}
\usage{
update_cache(force = FALSE, repo = options()$repos["CRAN"])
}
\arguments{
\item{force}{logical indicating whether to force an update}

\item{repo}{CRAN repo URL to use}
}
\description{
Downloads a CRAN packages file from the specified CRAN repo.
}
